requirements.txt
# This file contains the dependencies for the project.
# It is used to install the dependencies using pip.
# Example:
# requests==2.25.1
playwright==1.37.0
beautifulsoup4==4.12.2
PyYAML==6.0
google-generative-ai==0.3.1
jsonschema==4.17.3
ics==0.7.2